set API token to backend calls and ignore setting a token in interceptors if Authorization is given to the request
