+++ src/utils/interceptors.ts
@@ -2,23 +2,11 @@ import { IError, IErrorResponse, INotification } from '@/interfaces/common';
 import { clearLoading, setLoading } from '@/reducers/loaderSlice';
 import { notifyError } from '@/reducers/notificationSlice';
 import { AppStore } from '@/store';
-import { User } from 'oidc-client-ts';
 import axios, { AxiosError, AxiosResponse, InternalAxiosRequestConfig } from 'axios';
 
 let store: AppStore;
 
-const authority = 'https://tunnistus.test.hel.ninja/auth/realms/helsinki-tunnistus';
-const client_id = 'infraohjelmointi-ui-dev';
-
-const getUser = () => {
-  const oidcStorage = sessionStorage.getItem(`oidc.user:${authority}:${client_id}`);
-
-  if (!oidcStorage) {
-    return null;
-  }
-
-  return User.fromStorageString(oidcStorage);
-};
+const API_TOKEN_KEY = 'infraohjelmointi_api_token';
 
 // Add urls here that you don't want the interceptor to add a loader to
 const urlsToExclueFromLoading = [
@@ -77,10 +65,10 @@ const handleRequest = (req: InternalAxiosRequestConfig) => {
     store.dispatch(setLoading({ text: 'Loading request', id: req?.url ?? '' }));
   }
 
-  const user = getUser();
-  const token = user?.access_token;
+  const token = sessionStorage.getItem(API_TOKEN_KEY);
 
-  if (token) {
+  // Don't overwrite the token if it is given elsewhere
+  if (token && !req.headers.hasAuthorization()) {
     req.headers['Authorization'] = `Bearer ${token}`;
   }
